iot-monitoring-system/backend/requirements.txt
aniso8601
attrs
Click
flask
flask-restx
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-cors
itsdangerous
jinja2
MarkupSafe
mypy-extensions
pyrsistent
python-dateutil
pytz
six
sqlalchemy
werkzeug
pyjwt
pymysql